renamed services to servers
