+++ cmds/protocmd/main.go
@@ -18,7 +18,7 @@ import (
 	"github.com/maxthom/mir/internal/libs/boiler/mir_log"
 	"github.com/maxthom/mir/internal/libs/boiler/mir_signals"
 	"github.com/maxthom/mir/internal/libs/external/surreal"
-	"github.com/maxthom/mir/internal/services/protocmd_srv"
+	"github.com/maxthom/mir/internal/servers/protocmd_srv"
 	"github.com/maxthom/mir/pkgs/module/mir"
 	"github.com/rs/zerolog"
 	"golang.org/x/net/http2"
