+++ internal/servers/core_srv/readme.md
