+++ internal/servers/core_srv/server_integration_test.go
