+++ internal/servers/protocmd_srv/proto_test/buf.gen.yaml
