+++ internal/servers/protocmd_srv/proto_test/protocmd_test/v1/command.proto
