+++ internal/servers/protoflux_srv/proto_test/buf.gen.yaml
