+++ internal/servers/protoflux_srv/proto_test/protoflux_test/v1/telemetry.proto
