+++ internal/servers/protoflux_srv/server_integration_test.go
@@ -16,8 +16,8 @@ import (
 	"github.com/maxthom/mir/internal/libs/compression/zstd"
 	bus "github.com/maxthom/mir/internal/libs/external/natsio"
 	"github.com/maxthom/mir/internal/libs/test_utils"
-	"github.com/maxthom/mir/internal/services/core_srv"
-	protoflux_testv1 "github.com/maxthom/mir/internal/services/protoflux_srv/proto_test/gen/protoflux_test/v1"
+	"github.com/maxthom/mir/internal/servers/core_srv"
+	protoflux_testv1 "github.com/maxthom/mir/internal/servers/protoflux_srv/proto_test/gen/protoflux_test/v1"
 	core_apiv1 "github.com/maxthom/mir/pkgs/api/gen/proto/v1/core_api"
 	devicev1 "github.com/maxthom/mir/pkgs/device/gen/proto/mir/device/v1"
 	mirDevice "github.com/maxthom/mir/pkgs/device/mir"
