+++ internal/services/readme.md
@@ -1,0 +1,2 @@
+This folder is to put every services that
+are started from the main or that handle server requests.
